front/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
exa_py==1.0.9
requests==2.31.0
beautifulsoup4==4.12.2
anthropic==0.34.0 